update: exit status の取得が終了タイミングに依存しないように修正
waitpidにpidの指定を行っていないため、任意の子プロセスの終了を待つ。それ故、パイプ最後のプロセスの終了ステータスを確定で取得できるように、pidとexit statusの対応を取るように修正した。

Squashed commit of the following:

commit e06f46ee4fef6dd428f69f18b9e8601f47274da8
Author: GawinGowin <saraki@student.42tokyo.jp>
Date:   Sat Sep 14 11:08:13 2024 +0900

    fix

commit ee7195c9225a85db2f7b5b480a116582a42b1376
Author: GawinGowin <101625248+GawinGowin@users.noreply.github.com>
Date:   Sat Sep 14 01:57:04 2024 +0000

    tmp
