+++ src/exec/process_wait.c
@@ -1,75 +1,59 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   process_utils.c                                    :+:      :+:    :+:   */
+/*   process_wait.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: saraki <saraki@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/04/03 18:47:48 by saraki            #+#    #+#             */
-/*   Updated: 2024/08/29 21:13:22 by saraki           ###   ########.fr       */
+/*   Created: 2024/09/13 08:56:58 by saraki            #+#    #+#             */
+/*   Updated: 2024/09/14 11:04:45 by saraki           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "exec_int.h"
-#include "libft.h"
-#include "minishell.h"
 
+static int	cmp_pid(void *data, void *query_pt);
 static void	close_pipe_fds(t_pipelst *pipe_node);
-static int	pipe_fds(int *out_fd, int *in_fd);
 static int	check_status(int status);
-
-int	init_pipeline(t_pipelst *pipe_node)
-{
-	t_pipex	*pipe;
-	t_pipex	*next_pipe;
-
-	while (pipe_node->u_data.pipe_data != NULL)
-	{
-		pipe = (t_pipex *) pipe_node->u_data.pipe_data;
-		if (pipe_node->next->u_data.pipe_data != NULL)
-		{
-			next_pipe = pipe_node->next->u_data.pipe_data;
-			if (pipe_fds(&next_pipe->pipe_out_fd, &pipe->pipe_in_fd))
-				return (ERR);
-		}
-		pipe_node = pipe_node->next;
-	}
-	return (OK);
-}
-
-static int	pipe_fds(int *out_fd, int *in_fd)
-{
-	int	pipe_fd[2];
-
-	pipe_fd[0] = *out_fd;
-	pipe_fd[1] = *in_fd;
-	if (pipe(pipe_fd) < 0)
-		return (ERR);
-	*out_fd = pipe_fd[0];
-	*in_fd = pipe_fd[1];
-	return (OK);
-}
 
 int	wait_processes(t_pipelst *pipe_node)
 {
 	int			status;
+	pid_t		exited_pid;
+	t_pipelst	*exited_node;
 	int			err;
-	t_pipex		*data;
 
 	status = OK;
 	err = 0;
 	close_pipe_fds(pipe_node);
 	while (pipe_node->u_data.pipe_data != NULL)
 	{
-		data = pipe_node->u_data.pipe_data;
-		if (data->pids != 0 && waitpid(-1, &(data->exit_status), 0) == -1)
+		exited_pid = waitpid(-1, &status, 0);
+		if (exited_pid == -1)
+			err = GENERAL_ERR;
+		exited_node = (t_pipelst *)doub_lstsearch(
+				pipe_node->u_data.pipe_data->head_node, &exited_pid, cmp_pid);
+		if (exited_node->u_data.pipe_data != NULL)
+			exited_node->u_data.pipe_data->exit_status = status;
+		else
 			err = GENERAL_ERR;
 		pipe_node = pipe_node->next;
 	}
 	if (err != 0)
 		return (err);
-	status = pipe_node->next->u_data.pipe_data->exit_status;
-	return (check_status(status));
+	return (check_status(pipe_node->next->u_data.pipe_data->exit_status));
+}
+
+static int	cmp_pid(void *data, void *query_pt)
+{
+	pid_t		*pid;
+	t_pipex		*pipe_data;
+
+	pid = (pid_t *)query_pt;
+	pipe_data = (t_pipex *)data;
+	if (pipe_data->pids == *pid)
+		return (1);
+	return (0);
 }
 
 static void	close_pipe_fds(t_pipelst *pipe_node)
